PORTAL-2375: create immutable middleware options

+++ server/src/test/java/com/inventage/portal/gateway/proxy/middleware/authorization/passAuthorization/PassAuthorizationMiddlewareOptionsTest.java
@@ -71,7 +71,7 @@ public class PassAuthorizationMiddlewareOptionsTest {
 
         assertNotNull(options.getReconciliation());
         assertEquals(reconciliationEnabled, options.getReconciliation().isEnabled());
-        assertEquals(reconciliationIntervalMs, options.getReconciliation().getIntervalMs());
+        assertEquals(reconciliationIntervalMs.intValue(), options.getReconciliation().getIntervalMs());
 
         assertEquals(sessionScope, options.getSessionScope());
     }
