+++ server/src/test/java/com/inventage/portal/gateway/proxy/middleware/sessionLogoutFromBackchannel/BackChannelLogoutMiddlewareOptionsTest.java
@@ -69,6 +69,6 @@ public class BackChannelLogoutMiddlewareOptionsTest {
 
         assertNotNull(options.getReconciliation());
         assertEquals(reconciliationEnabled, options.getReconciliation().isEnabled());
-        assertEquals(reconciliationIntervalMs, options.getReconciliation().getIntervalMs());
+        assertEquals(reconciliationIntervalMs.intValue(), options.getReconciliation().getIntervalMs());
     }
 }
